requirements.txt
# pyspark==3.5.6
pymongo==4.6.1
psycopg2-binary==2.9.9
pandas==2.1.4
python-dateutil==2.8.2
apache-airflow-providers-postgres==5.8.0
apache-airflow-providers-apache-spark==4.10.0